+++ app/models/doctor.rb
@@ -1,4 +1,3 @@
 class Doctor < ApplicationRecord
-    has_many :reservations
+  has_many :reservations
 end
-
